Fix wrong locale id for require_login in admin
This fixes the wrong locale id introduced in https://git.io/fj0Yj.

The original and correct locale id was replaced a while ago with a wrong one.
That ended up forcing people to override the view or add the wrong id to their
locale files (even to en.yml ones). This fixes the issue by reverting the change.