Added verified purchaser
One feature that many other eCommerce sites have is the concept of
a "verified purchaser" which is when someone leaves a review for a
product they verifiably purchased. In the implementation here we
add a toggle on whether to display the verfied purchaser on the
frontend, a column on the database to store if the review is from
a verified purchaser, and logic when creating the review to set
that value.