Fix error when trying to call image.url
on the views it's belived that image is a `Paperclip::Attachment`
instance which can respond_to `#url` but in reality `image` is a
`Spree::Image` instance which can not respond_to `url`. So we need to use
the `#attachment` method instead to get the actual paperclip attachment
instance and pass it the size to get the actual url.