Change verified_purchaser to a before_save callback
Verified_purchaser being called on before_create would only
set the verfied_purchaser when a user creates a revew, obviously.

Recent updates have allowed users to update reviews, and it is possible
to create a review before purchasing the product, then come back
later after purchasing to update their review. Because of this
the before_create callback has been changed to before_save.
This way we can check if the user is a verified purchaser
before creating/updating a revew.