Note: This is a resubmission of PR #98 which appears to have been reverted.
When starting a Rails app using Solidus Reviews the following is logged
```
Reloading does not reboot the application, and therefore code executed during
initialization does not run again. So, if you reload Spree::ReviewsConfiguration, for example,
the expected changes won't be reflected in that stale Class object.
```

This change uses the same styles as solidus_frontend and solidus_backend to load the configuration class.